+++ Library/smartglove/Pose.h
@@ -17,15 +17,15 @@ class Pose
 {
 private:
 	vector<TolValue<double>> Articulation;
-	vector<TolValue<double>> Orientation;
+	vector<TolValue<unsigned short>> Orientation;
 
 public:
 	string name;
 	
 	Pose(string name, string data);
 	Pose(string name, vector<double> &ArticulationMeans, vector<double> &OrientationMeans);
-	Pose(string name, vector<double> &ArticulationMeans, vector<int> &OrientationMeans);
-	Pose(string name, vector<TolValue<double>> &Articulation, vector<TolValue<double>> &Orientation);
+	Pose(string name, vector<double> &ArticulationMeans, vector<unsigned short> &OrientationMeans);
+	Pose(string name, vector<TolValue<double>> &Articulation, vector<TolValue<unsigned short>> &Orientation);
 	~Pose();
 
 	bool operator==(Pose& curPose);
